test: use pytest for auto-download
